% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcSigClust.R
\name{CalcSigClust}
\alias{CalcSigClust}
\title{Perform SigClust tests pairwisely on the terminals of the dendrogram}
\usage{
CalcSigClust(seu, clusters)
}
\arguments{
\item{seu}{A Seurat object}

\item{clusters}{A vector of clustering labels}
}
\value{
A matrix of P values
}
\description{
Generate a dendrogram of cluster centroids on the left, with nodes indicating the significance of the P-values from the pairwise SigClust tests. On the right is a heatmap of P-values from the pairese SigClust tests.
}
