% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiagMultiKPlot.R
\name{DiagMultiKPlot}
\alias{DiagMultiKPlot}
\title{Make MultiK diagnostic plots}
\usage{
DiagMultiKPlot(ks, res)
}
\arguments{
\item{ks}{A vector of the number of clustering runs for each K}

\item{res}{A list of consensus matrix for each K}
}
\value{
A bar plot of the frequency of K, a plot of rPAC score for each K, and a scatterplot of (1-rPAC) vs. the frequency of K
}
\description{
Takes the output from MultiK() function, and generate diagnostic plots for determining optimal Ks
}
