% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClusters.R
\name{getClusters}
\alias{getClusters}
\title{Get clustering labels at optimal K}
\usage{
getClusters(seu, optK)
}
\arguments{
\item{seu}{A gene expression matrix}

\item{optK}{A vector of selected optimal Ks}
}
\value{
A matrix of clustering labels for each optimal K and the resolution parameters used to identifky the optimal K clusters.
}
\description{
Find the clustering labels at the selected optimal K levels.
}
