% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiK.R
\name{MultiK}
\alias{MultiK}
\title{MultiK main algorithm}
\usage{
MultiK(
  seu,
  resolution = seq(0.05, 2, 0.05),
  nPC = 30,
  reps = 100,
  pSample = 0.8,
  seed = NULL
)
}
\arguments{
\item{seu}{A Seurat object with normalized count}

\item{resolution}{A vector Seurat resolution parameters. Default is from 0.05 to 2 with step size of 0.05}

\item{nPC}{Number of principal components to use in clustering}

\item{reps}{Number of subsampling runs. Integer value. Default is 100}

\item{pSample}{Proportion of cells to sample. Numerical value. Default is 0.8}

\item{seed}{Optional numerical value. This sets a random seed for generating reproducible results}
}
\value{
A list with components: k is a vector of number of runs for each K. clusters is a list containing the clustering labels for each subsampling run at each resolution parameter. consensus is a list containing a consensus matrix for each K.
}
\description{
MultiK main algorithm: takes a preprocessed gene expression matrix as input. Then subsamples 80\% of the cells and applies standard Seurat pipeline on the subsampled data matrix 100 times over multiple resolution parameters.
}
