% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSigClust.R
\name{PlotSigClust}
\alias{PlotSigClust}
\title{Plot cluster centroids dendrogram and overlay P-values from pairwise SigClust tests}
\usage{
PlotSigClust(seu, clusters, pval)
}
\arguments{
\item{seu}{A Seurat object}

\item{clusters}{A vector of clustering labels}

\item{pval}{A matrix of P values from pairwise SigClust tests}
}
\value{
A diagnostic plot that shows the SigClust results
}
\description{
Generate a dendrogram of cluster centroids on the left, with nodes indicating the significance of the P-values from the pairwise SigClust tests. On the right is a heatmap of P-values from the pairese SigClust tests. The assignment of classes and subclasses is determined by the shape of the nodes. Classes are indicated by closed circles, and subclasses are indicated by open circles.
}
